Add science data processing excluding CDF attributes for now. Move sector rate subcommutation function to hit_l1a.py from decom_hit.py. Add global attrs for l1a science products to yaml file.
