+++ imap_processing/hit/l1a/hit_l1a.py
@@ -2,6 +2,7 @@
 
 import logging
 
+import numpy as np
 import xarray as xr
 
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
@@ -11,6 +12,7 @@ from imap_processing.hit.hit_utils import (
     get_datasets_by_apid,
     process_housekeeping_data,
 )
+from imap_processing.hit.l0.constants import MOD_10_MAPPING
 from imap_processing.hit.l0.decom_hit import decom_hit
 
 logger = logging.getLogger(__name__)
@@ -40,26 +42,122 @@ def hit_l1a(packet_file: str, data_version: str) -> list[xr.Dataset]:
     # Create the attribute manager for this data level
     attr_mgr = get_attribute_manager(data_version, "l1a")
 
+    l1a_datasets = []
+
     # Process l1a data products
     if HitAPID.HIT_HSKP in datasets_by_apid:
         logger.info("Creating HIT L1A housekeeping dataset")
-        datasets_by_apid[HitAPID.HIT_HSKP] = process_housekeeping_data(
-            datasets_by_apid[HitAPID.HIT_HSKP], attr_mgr, "imap_hit_l1a_hk"
+        l1a_datasets.append(
+            process_housekeeping_data(
+                datasets_by_apid[HitAPID.HIT_HSKP], attr_mgr, "imap_hit_l1a_hk"
+            )
         )
-
     if HitAPID.HIT_SCIENCE in datasets_by_apid:
-        # TODO complete science data processing
-        print("Skipping science data for now")
-        datasets_by_apid[HitAPID.HIT_SCIENCE] = process_science(
-            datasets_by_apid[HitAPID.HIT_SCIENCE], attr_mgr
+        l1a_datasets.extend(
+            process_science(datasets_by_apid[HitAPID.HIT_SCIENCE], attr_mgr)
+        )
+    return l1a_datasets
+
+
+def subcom_sectorates(sci_dataset: xr.Dataset) -> None:
+    """
+    Subcommutate sectorates data.
+
+    Sector rates data contains rates for 5 species and 10
+    energy ranges. This function subcommutates the sector
+    rates data by organizing the rates by species. Which
+    species and energy range the data belongs to is determined
+    by taking the mod 10 value of the corresponding header
+    minute count value in the dataset. A mapping of mod 10
+    values to species and energy ranges is provided in constants.py.
+
+    MOD_10_MAPPING = {
+        0: {"species": "H", "energy_min": 1.8, "energy_max": 3.6},
+        1: {"species": "H", "energy_min": 4, "energy_max": 6},
+        2: {"species": "H", "energy_min": 6, "energy_max": 10},
+        3: {"species": "4He", "energy_min": 4, "energy_max": 6},
+        ...
+        9: {"species": "Fe", "energy_min": 4, "energy_max": 12}}
+
+    The data is added to the dataset as new data fields named
+    according to their species. They have 4 dimensions: epoch
+    energy index, declination, and azimuth. The energy index
+    dimension is used to distinguish between the different energy
+    ranges the data belongs to. The energy min and max values for
+    each species are also added to the dataset as new data fields.
+
+    Parameters
+    ----------
+    sci_dataset : xr.Dataset
+        Xarray dataset containing parsed HIT science data.
+    """
+    # TODO:
+    #  - Update to use fill values defined in attribute manager which
+    #    isn't passed into this module nor defined for L1A sci data yet
+    #  - Determine naming convention for species data fields in dataset
+    #    (i.e. h, H, hydrogen, Hydrogen, etc.)
+    #  - consider moving this function to hit_l1a.py
+
+    # Calculate mod 10 values
+    hdr_min_count_mod_10 = sci_dataset.hdr_minute_cnt.values % 10
+
+    # Reference mod 10 mapping to initialize data structure for species and
+    # energy ranges and add 8x15 arrays with fill values for each science frame.
+    num_frames = len(hdr_min_count_mod_10)
+    data_by_species_and_energy_range = {
+        key: {**value, "rates": np.full((num_frames, 8, 15), fill_value=np.nan)}
+        for key, value in MOD_10_MAPPING.items()
+    }
+
+    # Update rates for science frames where data is available
+    for i, mod_10 in enumerate(hdr_min_count_mod_10):
+        data_by_species_and_energy_range[mod_10]["rates"][i] = sci_dataset[
+            "sectorates"
+        ].values[i]
+
+    # H has 3 energy ranges, 4He, CNO, NeMgSi have 2, and Fe has 1.
+    # Aggregate sector rates and energy min/max values for each species.
+    # First, initialize dictionaries to store rates and min/max energy values by species
+    data_by_species: dict = {
+        value["species"]: {"rates": [], "energy_min": [], "energy_max": []}
+        for value in data_by_species_and_energy_range.values()
+    }
+
+    for value in data_by_species_and_energy_range.values():
+        species = value["species"]
+        data_by_species[species]["rates"].append(value["rates"])
+        data_by_species[species]["energy_min"].append(value["energy_min"])
+        data_by_species[species]["energy_max"].append(value["energy_max"])
+
+    # Add sector rates by species to the dataset
+    for species, data in data_by_species.items():
+        # Rates data has shape: energy_index, epoch, declination, azimuth
+        # Convert rates to numpy array and transpose axes to get
+        # shape: epoch, energy_index, declination, azimuth
+        rates_data = np.transpose(np.array(data["rates"]), axes=(1, 0, 2, 3))
+
+        sci_dataset[species] = xr.DataArray(
+            data=rates_data,
+            dims=["epoch", f"{species}_energy_index", "declination", "azimuth"],
+            name=species,
+        )
+        sci_dataset[f"{species}_energy_min"] = xr.DataArray(
+            data=np.array(data["energy_min"]),
+            dims=[f"{species}_energy_index"],
+            name=f"{species}_energy_min",
+        )
+        sci_dataset[f"{species}_energy_max"] = xr.DataArray(
+            data=np.array(data["energy_max"]),
+            dims=[f"{species}_energy_index"],
+            name=f"{species}_energy_max",
         )
 
-    return list(datasets_by_apid.values())
 
-
-def process_science(dataset: xr.Dataset, attr_mgr: ImapCdfAttributes) -> xr.Dataset:
+def process_science(
+    dataset: xr.Dataset, attr_mgr: ImapCdfAttributes
+) -> list[xr.Dataset]:
     """
-    Will process science dataset for CDF product.
+    Will process science datasets for CDF products.
 
     Process binary science data for CDF creation. The data is
     grouped into science frames, decommutated and decompressed,
@@ -70,30 +168,70 @@ def process_science(dataset: xr.Dataset, attr_mgr: ImapCdfAttributes) -> xr.Data
     Parameters
     ----------
     dataset : xarray.Dataset
-        Dataset containing HIT science data.
+        A dataset containing HIT science data.
 
     attr_mgr : ImapCdfAttributes
         Attribute manager used to get the data product field's attributes.
 
     Returns
     -------
-    dataset : xarray.Dataset
-        An updated dataset ready for CDF conversion.
+    dataset : list
+        A list of science datasets ready for CDF conversion.
     """
     logger.info("Creating HIT L1A science datasets")
-
-    # Logical sources for the two products.
-    # logical_sources = ["imap_hit_l1a_count-rates", "imap_hit_l1a_pulse-height-event"]
 
     # Decommutate and decompress the science data
     sci_dataset = decom_hit(dataset)
 
-    # TODO: Complete this function
-    #  - split the science data into count rates and event datasets
-    #  - update dimensions and add attributes to the dataset and data arrays
-    #  - return list of two datasets (count rates and events)?
+    # Organize sector rates by species type
+    subcom_sectorates(sci_dataset)
 
-    # logger.info("HIT L1A event dataset created")
-    # logger.info("HIT L1A count rates dataset created")
+    # print(sci_dataset.data_vars)
+    # print(sci_dataset["sc_tick"])
+    # index = sci_dataset['sc_tick'].values.tolist().index(377990)
+    # print(index)
+    # print(sci_dataset['sc_tick'][index])
+    # print(sci_dataset['sc_tick'][:15])
+    # print(sci_dataset['livetime'].shape)
+    # print(sci_dataset['livetime'][:])
+    # print(sci_dataset['livetime'][7:])
+    # print(sci_dataset["H"].shape)
+    # print(sci_dataset["H"][0].shape)
+    # print(sci_dataset["H"][0])
 
-    return sci_dataset
+    # Split the science data into count rates and event datasets
+    # TODO: what else to include in pha? ccsds header info?
+    pha_raw_dataset = xr.Dataset(
+        {"pha_raw": sci_dataset["pha_raw"]}, coords={"epoch": sci_dataset["epoch"]}
+    )
+    count_rates_dataset = sci_dataset.drop_vars("pha_raw")
+
+    # Logical sources for the two products.
+    logical_sources = ["imap_hit_l1a_count-rates", "imap_hit_l1a_pulse-height-events"]
+
+    datasets = []
+    # Update attributes and dimensions
+    for dataset, logical_source in zip(
+        [count_rates_dataset, pha_raw_dataset], logical_sources
+    ):
+        dataset.attrs = attr_mgr.get_global_attributes(logical_source)
+        # TODO: add CDF attributes once they're defined for L1A science data
+
+        # # Assign attributes and dimensions to each data array in the Dataset
+        # for field in dataset.data_vars.keys():
+        #     # Create a dict of dimensions using the DEPEND_I keys in the
+        #     # attributes
+        #     dims = {
+        #         key: value
+        #         for key, value in attr_mgr.get_variable_attributes(field).items()
+        #         if "DEPEND" in key
+        #     }
+        #     dataset[field].attrs = attr_mgr.get_variable_attributes(field)
+        #     dataset[field].assign_coords(dims)
+        #
+        dataset.epoch.attrs = attr_mgr.get_variable_attributes("epoch")
+        datasets.append(dataset)
+
+        logger.info(f"HIT L1A dataset created for {logical_source}")
+
+    return datasets
